split hapi basic file into 2 with a shared implementation
